requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via insightweaver (pyproject.toml)
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.40.0
    # via insightweaver (pyproject.toml)
anyio==4.12.0
    # via
    #   anthropic
    #   httpx
attrs==25.4.0
    # via aiohttp
beautifulsoup4==4.12.3
    # via
    #   insightweaver (pyproject.toml)
    #   markdownify
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via nltk
distro==1.9.0
    # via anthropic
feedparser==6.0.11
    # via insightweaver (pyproject.toml)
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   insightweaver (pyproject.toml)
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.12.0
    # via anthropic
joblib==1.5.3
    # via nltk
lxml==6.0.2
    # via insightweaver (pyproject.toml)
markdown-it-py==4.0.0
    # via rich
markdownify==1.2.2
    # via insightweaver (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
nltk==3.9.2
    # via textblob
pillow==12.0.0
    # via insightweaver (pyproject.toml)
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.12.5
    # via
    #   anthropic
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via insightweaver (pyproject.toml)
pygments==2.19.2
    # via rich
python-dotenv==1.0.1
    # via
    #   insightweaver (pyproject.toml)
    #   pydantic-settings
pyyaml==6.0.2
    # via insightweaver (pyproject.toml)
regex==2025.11.3
    # via nltk
requests==2.32.5
    # via insightweaver (pyproject.toml)
rich==14.2.0
    # via insightweaver (pyproject.toml)
sgmllib3k==1.0.0
    # via feedparser
six==1.17.0
    # via markdownify
sniffio==1.3.1
    # via anthropic
soupsieve==2.8.1
    # via beautifulsoup4
sqlalchemy==2.0.45
    # via insightweaver (pyproject.toml)
textblob==0.19.0
    # via insightweaver (pyproject.toml)
tqdm==4.67.1
    # via nltk
typing-extensions==4.15.0
    # via
    #   anthropic
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.2
    # via requests
yarl==1.22.0
    # via aiohttp
